demonstrator/backend/requirements.txt
# requirements.txt - Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
librosa==0.10.1
soundfile==0.12.1
opensmile==2.5.0
torch==2.0.1
transformers==4.35.0
TTS==0.22.0
parselmouth-praat==0.4.3
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
psycopg2-binary==2.9.9
redis==5.0.1
